requirements.txt
Django==2.1.3
djangorestframework==3.9.0
drf-nested-routers==0.91.0
